complete ssi process and reject counterparty request

+++ projects/cob/src/main/java/com/noble/bpm/cob/subprocess/CancelProcessDelegate.java
@@ -1,8 +1,11 @@
 package com.noble.bpm.cob.subprocess;
 
+import java.util.List;
+
 import org.camunda.bpm.engine.RuntimeService;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
 import org.camunda.bpm.engine.delegate.JavaDelegate;
+import org.camunda.bpm.engine.runtime.Execution;
 
 public class CancelProcessDelegate implements JavaDelegate {
 
@@ -10,8 +13,11 @@ public class CancelProcessDelegate implements JavaDelegate {
 	public void execute(DelegateExecution execution) throws Exception {
 		String subProcessInstanceId = (String) execution.getVariable("ssiApprovalInstanceId");
 		RuntimeService runtimeService = execution.getProcessEngineServices().getRuntimeService();
-		runtimeService.setVariable(subProcessInstanceId, "cancelReason", "original counterparty request is rejected");
-		runtimeService.deleteProcessInstance(subProcessInstanceId, "original counterparty request is rejected");
+		List<Execution> subProcessInstances = runtimeService.createExecutionQuery().executionId(subProcessInstanceId).list();
+		if (subProcessInstances.isEmpty() == false) {
+			runtimeService.setVariable(subProcessInstanceId, "cancelReason", "original counterparty request is rejected");
+			runtimeService.deleteProcessInstance(subProcessInstanceId, "original counterparty request is rejected");
+		}
 	}
 
 }
